% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compares test MSE between different sets of predictors}
\usage{
compare_models(dat, response_var, pred_var_1, pred_var_2)
}
\arguments{
\item{dat}{A dataset}

\item{response_var}{An unquoted variable name}

\item{pred_var_1}{An unquoted variable name}

\item{pred_var_2}{An unquoted variable name}
}
\value{
A data frame
}
\description{
Compares test MSE between different sets of predictors
}
