% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_betas.R
\name{compute_betas_once}
\alias{compute_betas_once}
\title{Calculates the coefficients of a multiple regression using matrix decomposition}
\usage{
compute_betas_once(y, X)
}
\arguments{
\item{y}{A n by 1 matrix of response variable values}

\item{X}{A n by p matrix of predictor values}
}
\value{
A data frame of coefficient estimates
}
\description{
Calculates the coefficients of a multiple regression using matrix decomposition
}
