% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{get_MSE}
\alias{get_MSE}
\title{Computes MSE from predicted values}
\usage{
get_MSE(results, truth, estimate)
}
\arguments{
\item{results}{A data frame of predictions and truths}

\item{truth}{The name of the observed values column}

\item{estimate}{The name of the predicted values column}
}
\value{
The MSE
}
\description{
This function takes a data frame of predictions in the form outputted by
the \code{predict_from_coefs} function.
}
