% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_betas.R
\name{make_betas}
\alias{make_betas}
\title{Calculates the coefficients of a multiple regression for different predictor combos.}
\usage{
make_betas(dat, response_var, pred_var_1, pred_var_2)
}
\arguments{
\item{dat}{A dataset}

\item{response_var}{An unquoted variable name}

\item{pred_var_1}{An unquoted variable name}

\item{pred_var_2}{An unquoted variable name}
}
\value{
A data frame
}
\description{
Calculates the coefficients of a multiple regression for different predictor combos.
}
