% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{viz_cat_pred}
\alias{viz_cat_pred}
\title{Makes a basic barplot comparing a categorical predictor to a response variable.}
\usage{
viz_cat_pred(dat, response_var, pred_var)
}
\arguments{
\item{dat}{A dataset}

\item{response_var}{An unquoted variable name}

\item{pred_var}{An unquoted variable name}
}
\description{
Makes a basic barplot comparing a categorical predictor to a response variable.
}
