% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{viz_quant_pred}
\alias{viz_quant_pred}
\title{Makes a basic scatterplot comparing a quantitative predictor to a response variable.}
\usage{
viz_quant_pred(dat, response_var, pred_var)
}
\arguments{
\item{dat}{A dataset}

\item{response_var}{An unquoted variable name}

\item{pred_var}{An unquoted variable name}
}
\description{
Makes a basic scatterplot comparing a quantitative predictor to a response variable.
}
